requirements.txt
asgiref==3.3.1
boto3==1.17.3
botocore==1.20.3
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.3
django-allauth==0.44.0
django-filter==2.4.0
django-on-heroku==1.1.2
django-phonenumber-field==5.0.0
django-storages==1.11.1
django-taggit==1.3.0
django-widget-tweaks==1.4.8
gunicorn==20.0.4
idna==2.10
install==1.3.4
jmespath==0.10.0
oauthlib==3.1.0
phonenumbers==8.12.13
Pillow==8.0.1
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.0
pyp==2.12
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
